maptodouble insted

+++ src/FindExtreme.java
@@ -1,4 +1,4 @@
 @FunctionalInterface
 public interface FindExtreme {
-    int maxOrMin(int a, int b);
+    double maxOrMin(double a, double b);
 }
